ai-log-monitor/requirements.txt
pyyaml
scikit-learn
watchdog
pandas
streamlit

